+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- added: Filecoin FEVM token support
 - added: (EVM) Accept transactions with a provided `otherParams.nonce` even if pending transactions exist
 
 ## 4.55.8 (2025-08-01)